The pdftops filter did not compile with GCC 4.0 (STR #1226)
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@4558 7a7537e8-13f0-0310-91df-b6672ffda945

+++ pdftops/GfxFont.cxx
@@ -1008,7 +1008,7 @@ Gushort *Gfx8BitFont::getCodeToGIDMap(FoFiTrueType *ff) {
   // try the TrueType 'post' table to handle any unmapped characters
   for (i = 0; i < 256; ++i) {
     if (!map[i] && (charName = enc[i])) {
-      map[i] = (Gushort)(int)ff->mapNameToGID(charName);
+      map[i] = (Gushort)(long)ff->mapNameToGID(charName);
     }
   }
